cleanup

+++ src/test/java/clipper2/TestLines.java
@@ -1,6 +1,12 @@
 package clipper2;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -8,13 +14,10 @@ import org.junit.jupiter.params.provider.MethodSource;
 
 import clipper2.ClipperFileIO.TestCase;
 import clipper2.core.Point64;
-
-import java.io.IOException;
-import java.util.*;
-import java.util.stream.Collectors;
+import clipper2.engine.Clipper64;
 
 class TestLines {
-	
+
 	private static final List<Arguments> testCases() throws IOException {
 		return ClipperFileIO.loadTestCases("Lines.txt").stream().map(t -> Arguments.of(t, t.caption(), t.clipType(), t.fillRule()))
 				.collect(Collectors.toList());
@@ -38,7 +41,7 @@ class TestLines {
 		}
 
 		if (test.count() > 0 && Math.abs(solution.size() - test.count()) > 2
-				&& (double) Math.abs(solution.size() - test.count()) / test.count() > 0.02) {
+				&& Math.abs(solution.size() - test.count()) / test.count() > 0.02) {
 			assertTrue(Math.abs(solution.size() - test.count()) < 4, String.format("Incorrect count in test %1$s", test.caption()));
 		}
 	}
